コメントの追記とprofileをauth内に移動
